医药翻译-编码目录/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "medical-translate-backend"
version = "0.1.0"
description = "Backend scaffold for the medical translation product"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "greenlet>=3.0,<4.0",
  "aiosqlite>=0.20,<1.0",
  "alembic>=1.13,<2.0",
  "pydantic>=2.8,<3.0",
  "pydantic-settings>=2.3,<3.0",
  "email-validator>=2.2,<3.0",
  "python-multipart>=0.0.9,<1.0",
  "python-jose[cryptography]>=3.3,<4.0",
  "passlib[bcrypt]>=1.7,<2.0",
  "httpx>=0.27,<1.0",
  "reportlab>=4.0,<5.0",
  "pypdf>=6.0,<7.0",
  "python-docx>=1.1,<2.0",
  "structlog>=24.2,<25.0",
  "redis>=5.0,<6.0",
  "celery>=5.4,<6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2,<9.0",
  "pytest-asyncio>=0.23,<1.0",
  "ruff>=0.5,<1.0",
  "mypy>=1.10,<2.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = ["app*"]
